+++ src/reactionsystem.jl
@@ -1161,6 +1161,36 @@ function dependants(rx, network)
     dependents(rx, network)
 end
 
+"""
+isautonomous(rs::ReactionSystem)
+
+Checks if a system is autonomous (i.e. no rate or equation depend on the independent variable(s)).
+Example:
+```julia
+rs1 = @reaction_system
+    (p,d), 0 <--> X
+end
+isautonomous(rs1) # Returns `true`.
+
+rs2 = @reaction_system
+    (p/t,d), 0 <--> X
+end
+isautonomous(rs2) # Returns `false`.
+```
+"""
+function isautonomous(rs::ReactionSystem)
+    # Get all variables occurring in reactions and equations.
+    vars = Set()
+    for eq in equations(rs)
+        (eq isa Reaction) ? get_variables!(vars, eq.rate) : get_variables!(vars, eq)
+    end
+
+    # Checks for iv and spatial ivs
+    (get_iv(rs) in vars) && return false
+    any(in(vars), get_sivs(rs)) && return false
+    return true
+end
+
 
 ### `ReactionSystem` Remaking ###